GenAI/campusX/pyproject.toml
[project]
name = "campusx"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "faiss-cpu>=1.11.0",
    "google-generativeai>=0.8.5",
    "huggingface-hub>=0.33.1",
    "langchain>=0.3.26",
    "langchain-anthropic>=0.3.16",
    "langchain-community>=0.3.26",
    "langchain-core>=0.3.66",
    "langchain-google-genai>=2.0.10",
    "langchain-huggingface>=0.3.0",
    "langchain-openai>=0.3.26",
    "numpy>=2.3.1",
    "openai>=1.92.2",
    "python-dotenv>=1.1.1",
    "scikit-learn>=1.7.0",
    "sentence-transformers>=4.1.0",
    "transformers>=4.53.0",
    "youtube-transcript-api>=1.1.0",
]
